17/03/2025 #2
Reworked the movement system to use an array of Cells instead of a single (X,Z) destination coordinate.
I had to overwrite the packets, the utility functions, the player data, the game state data, and how they all connect, also had to update the player spawn logic.
In the godot client, I had to completely override the movement logic to use the new array instead of a direct X and Z coordinates.
Now in the client, its interpolating between the current cell and the target cell, sliding on the floor.

TO DO:
- I still need to work on the look at rotation.
- I need to work on the interpolation speed based on the number of cells to traverse in a single tick and the player's speed from the server.