20/03/2025 #2
- Fixed the movement bug, I had to ask AI for help with the overlapping cells at the start of the second path with the cells at the end of our current path.
AI came up with a loop that iterates both and eliminates the overlap with a slice at the end.
Then I had to fix some of the logic in my code, I was changing the next cell every time I called update_destination(), and I only had to update the next cell if my character was idle.

Now that the movement bugs are fixed:
- I have to work in the rotation per cell.
- I have to fix the double login issue.
- I found a new problem, the heartbeat latency packet no longer works with the packet rate limit, so I have to overhaul that part of the code.