25/04/2025 #2
- Reworked both back end and front end code to send a single position with the update destination packet, and have the client do the pathfinding for both local and remote players.
- Fixed the movement it feels better, it doesn't freeze or fails as much, but I still have a bug.

BUG:
When I move diagonally, the character slides really quickly to the grid_position he was last tick, and then starts from there, instead of trying to go back from where he currently is, making it very jarring. I have to fix this before implementing A* in the client.. because the bug was going to be problematic to track down later.