09/06/2025 #3
- FIXED: When autopiloting to the new server position, the character stands idle while in move state, animating in place for a tick before rotating to the next position, I think a cell position is being repeated in the path somewhere, thats why it hangs there in place.
I fixed this bug by slicing the first cell from the path, preventing the walk in place animation hang for a tick.
- BUG: When a character (due to latency or pathfinding) ends up in the same position as another player, it causes a desync locally yet remotely its fine.
FIXED: I fixed it partly, when clicking somewhere else, the character might path find correctly, or sometimes it caused a desync with the server, so the server will correct that client but it was causing a weird sliding bug, I fixed it by doing an intentional teleport to the accurate position right away. Created a function called _teleport_to_position() to achieve this.
I should probably check my local grid on every immediate_grid_destination too, to try and pathfinding another way to my grid_destination if an obstacle got in the way.